Create B 110 : The Continue Statement.c
Counter keyword to the break statement is the continue statement, The break statement stops the execution of the loop while the continue statement takes it to the next iteration.

Whenever a continue statement is invoked, no other statements after it will be executed for the iteration and the iteration ends. The execution of the next iteration begins at the beginning of the loop body once again.

For this task, write a program to input one number N, and to print all the natural numbers until N except those that are a multiple of 3.