Create B 114 : Nested For Loops.c
It is possible in C to write one for loop inside another. This is known as nested for loop. For example :

for (i = 0; i < n; i++)
{
	for (j = 0; j < n; j++)
	{
		//block of statements.
	}
}
To complete this task write a program that takes a number N as input and prints a square block of asterisks '*' of size NxN.